[TC] Regenerated demo for tc_tc_capture_mode_sam_e70_xult
MH3-67827

+++ apps/tc/tc_capture_mode/firmware/src/config/sam_e70_xult/peripheral/tc/plib_tc1.c
@@ -120,7 +120,7 @@ void TC1_CH2_ComparePeriodSet (uint16_t period)
 /* Read the period value */
 uint16_t TC1_CH2_ComparePeriodGet (void)
 {
-    return TC1_REGS->TC_CHANNEL[2].TC_RC;
+    return (uint16_t)TC1_REGS->TC_CHANNEL[2].TC_RC;
 }
 
 /* Set the compare A value */
